[Refactor] 반환 형식 수정
ResponseEntity로 body에 ResponseResult 타입으로 결과를 반환하도록 수정
